Added finalPosition collider to movement, sprite sorter, & update tools.
The finalPosition collider is a CircleCollider2D that every moveable character object will need. This collider is cast out to the final position the character is moving and "saves" the spot for the moving character, preventing another character from moving into the same tile.
The sprite sorting script is one that I've used on previous projects. It dynamically updates the order in the sprite sorting layer depending on where the object is located on the x-y plane.
There were also minor tweaks to tools and code pieces to accommodate these changes.