Combat tile distance check. Small cleanup.
Added ability to check the distance between two points in 2D space, returning the number of "tiles" the distance corresponds to. This will allow systems to perform tasks that require knowledge of distance in tile-space.
The cleanup centered around commented code that was no longer needed now that the new systems have been tested.