Projectiles work. Temp fix for Spawner collisions.
Projectiles can be instantiated by an actor and fired in a specified direction.
This took some trickery with the quaternions and angle of the velocity since the sprite is not facing the proper "up" direction (this as so I could make arrow sprites within a 16x16 tile that had a sharp point).
The trigger collider for the enemy spawners are being detected by other game objects, so collisions with them are being ignored for now. Unfortunately, this means that functionality of the collider (detecting when it is not blocked by another object) is unable to perform its job. Due to this, a more permanent solution needs to be found.