Debugging TurnSystem.
TurnSystem is having issues that stem from it not being a MonoBehaviour. It is unable to check (by default) if a GameObject or MonoBehaviour script has been Destroy()ed. Attempts to caste the elements within the actor list seem to show some progress, but it still does not work. Issue is not with the List.RemoveAll method. All testing has been done with Projectiles up to this point.