Fixed bug with Enemy health. Found issue with inputEnabled (needs fix).
The CheckAlive abstract method for the Enemy class was too clunky. When in combat, the calls would happen ON the enemy's turn and not when the damage was taken (so the enemy would get a turn after taking fatal damage, though the turn would end immediately, but not before other objects interacted with it). By moving that check into the enemy's TakeDamage method on the base class, the GameObject is now destroyed immediately.
There is still an issue with the user input being disabled. This happens when the player kills the last enemy, then leaves combat, but keeps spamming arrows (haven't tested with other types of attacks). Upon doing that, usually (but not always) the game will keep running, but the user input variable is still set to false, so the user cannot make actions. Though found, I still don't know why this is happening, but the problem will have to wait for tomorrow.