Fixed input disable issue (I think). Sped up movement.
The input delay issue is fixed, I am fairly positive. If the player's coroutine that disables input is run when not in combat, the input will not be disabled. The one issue with this is if the CombatOver event is not triggered. In that case, the input will still get disabled. My stance on that for right now is: if that event isn't firing, then there is a bigger issue than the user input disabling.
The movement for characters (the time for the rigidbodies to move from spot to the other in their coroutines) has been sped up and the user's input delay has been reduced. These are settings that will probably continue to be fiddled with until the perfect "feel" is met.