Fixed projectile movement.
What seemed like the biggest issue in the code (and one I thought was going to cause a lot of pain and suffering) ended up being a one line fix. Not only that, but the fix was exactly what I thought it should be. By stopping and figuring out the vector math, before just diving in, I came up with a solution and it was correct.
Relishing this moment, as they are not common.