Multiple enemies at once works. Need to simplify to FSM for enemies.
Multiple enemies can be in a scene at once and combat works. Next goal is to implement a streamlined FSM for the state of the enemy (Not in combat, but TurnSystem active; In combat and TurnSystem active; not in combat and turn system not active). The states could be simplified (for example, if in combat, the turn system has to be active).