Current bug - UserInput disabled again. Moved enemy behav. to base class
Seems to happen with more enemies spawned (about 5+). Changed the Enemy base class to take on some of the functionality that should be shared by all enemies that use the turn system. It is not clear yet if the current bug is due to that change or not. Will try to take out that change next and determine if the bug still exists or not.