Fixed issue with game loop getting stuck.
Looks like the issue was with objects in the TurnSystem (projectiles / enemies / etc.) getting Destroy()ed during their turn. When they would get destroyed, the coroutines for the behavior would stop, but nothing would notify the TurnSystem that it needs to move onto the next object in the queue because the object was destroyed before it could raise its "turn over" event.
By setting a flag that tracks when it is the object's turn and then, when it gets disabled, if it is its turn, it signals that its turn is over.
This appears to work. After a long test-session, there were no hang-ups produced or input disabled issues.
Next, with the changes all in place, a heavy refactoring needs to be done.