Updated projectiles to take range value. Added thrown poison bottles.
Instead of being an AOE attack, thrown destructibles are now a type of projectile, which makes a lot more sense in hindsight. To do this, projectiles had to be altered to take a Range value, which required their starting point to be saved upon creation so that their current traveled distance could be calculated.
With a quick implementation, everything works as expected. Will keep an eye on any bugs that may pop up later, but nothing amiss as of now.