Fixed bugs: Projectile spawn on player; Poison Puddles not working.
Fixed an issue where projectiles could be spawned on top of the player, thus dealing damage to themselves.
Poison puddles that spawned from the thrown Bottles were not dealing damage and were not ticking their duration correctly, but both of those things have been corrected.
Added a sprite for a single-tile poison puddle for testing out a smaller version. Tweaks need to be made to the AoE durations as well - 3 seconds for the Poison is not very long.

Remaining issue: Projectiles can deal double damage to enemies when striking them. It looks like this happens when the projectile hits the enemy from a diagonal trajectory (hitting on the corner).