Added third enemy. Updated life bar. Fixed bugs.
Bugs discovered were the enemies (except Wisps) had their rigidbodies set to kinematic instead of dynamic. This caused issues where the enemies were able to move through walls and didn't react to physics. Life bar was updated, but the design was changed from the initial one due to the size of the first design. Voidseer enemy was added and can target the player with ranged (and melee) attacks.
Current issue is that enemies can continuously attack even when the time is "stopped."