Player attack delay and collides w/enemies. Enemies flash red when hit.
Implemented an attack delay for the player so that they cannot spam attacks (mainly concerned with the ranged attacks, which may still need some kind of ammo).
Player now collides with enemies (doesn't move through them) - this is meant to make melee attacking easier for the player and to reduce the amount of sorting needing to be done for the sprites.
Enemies now have their sprites change to a reddish color when taking damage (as a visual indicator). This still needs to be implemented for the player, as well as an audio indicator.