Level 1 has had its size reduced. CDs added to ranged&thrown.
Some of the rooms in the first level have had their sizes reduced. Making rooms on the smaller-side seems to make for better flow (not as much running around).
A cooldown was added to the player's ranged attack (they can still attack, but if they don't wait then they'll take damage).
Cooldown has been added to the player's thrown ability that disables the ability for a set amount of time.