Tweaked enemy values. Fixed null when premature end. Win works and death
Tweaked some of the health and damage values for enemies.
There were null references that would occur with the activated room system when the scene was ended prematurely. Didn't break anything, but took care of them nonetheless.
Killing all enemies in the scene now completes the level.
Dying now disables the player object (the scene ending was already in place).